requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
gunicorn==22.0.0
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
